+++ src/Encamina.Enmarcha.Email.Abstractions/EmailAttachmentSpecification.cs
@@ -15,7 +15,7 @@ public class EmailAttachmentSpecification
     /// <summary>
     /// Gets the <see cref="ContentType">content type</see> of the attachment in this specification.
     /// </summary>
-    public ContentType ContentType { get; init; }
+    public ContentType? ContentType { get; init; }
 
     /// <summary>
     /// Gets the binary data of the attachment in this specification.