+++ src/Encamina.Enmarcha.Email.Abstractions/SmtpClientOptions.cs
@@ -6,11 +6,11 @@ using Encamina.Enmarcha.Entities.Abstractions;
 namespace Encamina.Enmarcha.Email.Abstractions;
 
 /// <summary>
-/// Represent's common SMTP configuration parameters or options.
+/// Represents common SMTP configuration parameters or options.
 /// </summary>
 public class SmtpClientOptions : INameable
 {
-    private string name = null;
+    private string? name = null;
 
     /// <summary>
     /// Gets or sets the host name of the SMTP service.