+++ src/Encamina.Enmarcha.Email.MailKit/EmailService.cs
@@ -50,7 +50,7 @@ internal sealed class EmailService : IEmailBuilder, IEmailProvider, ISmtpClientO
         => AddAttachment(fileName, data, string.IsNullOrWhiteSpace(contentTypeValue) ? null : new ContentType(contentTypeValue));
 
     /// <inheritdoc/>
-    public IEmailBuilder AddAttachment(string fileName, byte[] data, ContentType contentType)
+    public IEmailBuilder AddAttachment(string fileName, byte[] data, ContentType? contentType)
     {
         Specification.Attachments.Add(new EmailAttachmentSpecification()
         {
@@ -63,7 +63,7 @@ internal sealed class EmailService : IEmailBuilder, IEmailProvider, ISmtpClientO
     }
 
     /// <inheritdoc/>
-    public IEmailBuilder AddRecipient(string emailAddress, string recipientName = null, EmailRecipientType recipientType = EmailRecipientType.TO)
+    public IEmailBuilder AddRecipient(string emailAddress, string? recipientName = null, EmailRecipientType recipientType = EmailRecipientType.TO)
     {
         Guard.IsNotNullOrWhiteSpace(emailAddress);
 
@@ -114,7 +114,7 @@ internal sealed class EmailService : IEmailBuilder, IEmailProvider, ISmtpClientO
     }
 
     /// <inheritdoc/>
-    public IEmailBuilder SetBody(string body, bool isHtml = false)
+    public IEmailBuilder SetBody(string? body, bool isHtml = false)
     {
         Specification.Body = body;
         Specification.IsHtmlBody = isHtml;
@@ -128,7 +128,7 @@ internal sealed class EmailService : IEmailBuilder, IEmailProvider, ISmtpClientO
     public IEmailBuilder SetSender(string emailAddress) => SetSender(emailAddress, null);
 
     /// <inheritdoc/>
-    public IEmailBuilder SetSender(string emailAddress, string senderName)
+    public IEmailBuilder SetSender(string emailAddress, string? senderName)
     {
         Guard.IsNotNullOrWhiteSpace(emailAddress);
         Guard.IsTrue(emailAddress.IsValidEmail(), nameof(emailAddress), @"Parameter is not a valid e-mail format!");
@@ -143,7 +143,7 @@ internal sealed class EmailService : IEmailBuilder, IEmailProvider, ISmtpClientO
     }
 
     /// <inheritdoc/>
-    public IEmailBuilder SetDefaultSender(string senderName = null) => SetSender(SmtpClientOptions.User, senderName);
+    public IEmailBuilder SetDefaultSender(string? senderName = null) => SetSender(SmtpClientOptions.User, senderName);
 
     /// <inheritdoc/>
     /// <remarks>This implementation does not allows the subject to be <see langword="null"/>.</remarks>