+++ CHANGELOG.md
@@ -83,6 +83,7 @@ Previous classification is not required if changes are simple or all belong to t
   - `Encamina.Enmarcha.Bot`
   - `Encamina.Enmarcha.Core`
   - `Encamina.Enmarcha.Data`
+  - `Encamina.Enmarcha.Email`
 - Corrected a typo in the Spanish error message in `ResponseMessages.es.resx` from "ha encontrar" to "ha encontrado".
  
 ## [8.1.5]